requirements.txt
-i https://pypi.org/simple
grpcio==1.59.0; python_version >= '3.7'
grpcio-tools==1.59.0; python_version >= '3.7'
kaikosdk==1.25.0
protobuf==4.21.12; python_version >= '3.7'
retrying==1.3.3
setuptools==75.2.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
requests
